requirements.txt
flask==3.0.0
pynacl==1.5.0
upnpy==1.1.8
cryptography==41.0.7
python-dotenv==1.0.0 